+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 0.22
 - new function `dl_newinsight` to create READ ACCESS DB read token in table `datalink.insight`  
 - new function `dl_lco(datalink)` to get LCO for a (linked) datalink
 - renamed `datalink.dl_url` to `datalink.url`
+- remove read_access flag in datalink (r)
 
 Version 0.21
 ------------